[BTAT-3861] Refactored UPdateVatSubscriptionService into Seperate Services
